+++ doc/changelog.rst
@@ -12,6 +12,11 @@ General
   for the optimization. This allows to decrease the number of initial points needed for finding the contacts and is more
   robust for very small or large geom sizes.
 
+Bug fixes
+^^^^^^^^^
+- Fix bug in Cartesian actuation with movable refsite, as when using body-centric Cartesian actuators on a quadruped.
+  Before this fix such actuators could lead to non-conservation of momentum.
+
 Version 3.0.1 (November 15, 2023)
 ---------------------------------